requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-jazzmin==2.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0
libvirt-python==8.3.0
MouseInfo==0.1.3
Pillow==9.1.1
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
python3-xlib==0.15
pytweening==1.0.4
sqlparse==0.4.2
whitenoise==6.1.0
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-jazzmin==2.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0
libvirt-python==8.3.0
MouseInfo==0.1.3
Pillow==9.1.1
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
python3-xlib==0.15
pytweening==1.0.4
sqlparse==0.4.2
whitenoise==6.1.0
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-jazzmin==2.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0
libvirt-python==8.3.0
MouseInfo==0.1.3
Pillow==9.1.1
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
python3-xlib==0.15
pytweening==1.0.4
sqlparse==0.4.2
whitenoise==6.1.0
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-jazzmin==2.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0
libvirt-python==8.3.0
MouseInfo==0.1.3
Pillow==9.1.1
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
python3-xlib==0.15
pytweening==1.0.4
sqlparse==0.4.2
whitenoise==6.1.0
